# PR #5: Performance Test Optimization & Edge Case Fixes